pyproject.toml
[project]
name = "raiflowv2-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AndreRichartz", email = "andrerichartz07@gmail.com"},
]
dependencies = ["django>=5.2.6", "djangorestframework>=3.16.1", "django-cors-headers>=4.9.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
